pyproject.toml
[tool.poetry]
name = "backend-"
version = "0.1.0"
description = ""
authors = ["Ahlemmer <ahlemmerabtene@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-filter = "^24.3"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
django-environ = "^0.11.2"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"
pytest = "^8.3.3"
pytest-xdist = "^3.6.1"
pytest-django = "^4.9.0"
model-bakery = "^1.20.0"
factory-boy = "^3.3.1"
django-redis = "^5.4.0"
celery = "^5.4.0"
redis = "^5.2.0"
kombu = "^5.4.2"

[tool.poetry.dev-dependencies]
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
colorlog = "^6.8.2"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "erp.core.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["erp"]
pythonpath = ["."]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
